requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
backtrader==1.9.78.123
blinker==1.9.0
cachetools==5.5.1
ccxt==4.4.50
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
fonttools==4.55.5
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
dependency_injector>=4.0.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
multidict==6.1.0
narwhals==1.23.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.2
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybit==5.8.0
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.41.1
ta-lib @ file:///D:/Programa%C3%A7%C3%A3o/Bybit_Watcher/assets/ta_lib-0.6.0-cp313-cp313-win_amd64.whl#sha256=926cc1b6e18a1aa8b18274d4d75667bfe8f3c0bb8a6e32e91d927f076322af3f
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
tweepy==4.15.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==3.0.6
win32_setctime==1.2.0
yarl==1.18.3
zipp==3.21.0
